requirements.txt
Flask==2.0.1
gunicorn==20.1.0
rdkit-pypi
numpy
nltk
